Add "View reasoning" to the per-application policy editor too
The reasoning panel (#291) was only wired into the global routing policy editor,
but policies are just as often generated per application (Applications → app →
Routing policy), and that flow returns the same evidence payload. Extract
ReasoningDrawer into a shared components/ReasoningDrawer.jsx and use it in both
editors so "View reasoning" appears wherever a policy is generated.

- New components/ReasoningDrawer.jsx (moved verbatim from Routing.jsx).
- Routing.jsx imports it (drops the now-unused local copy + Drawer import).
- ApplicationDetail.jsx captures result.evidence from generateAppPolicy and
  shows the same "View reasoning" slide-over.
