+++ web/src/components/ReasoningDrawer.jsx
@@ -1,0 +1,92 @@
+import React from "react";
+import { fmt } from "../api.js";
+import { Drawer, Badge } from "./ui.jsx";
+
+const CONFIDENCE_TONE = { high: "green", medium: "amber", low: "gray" };
+
+// Slide-over that explains WHY the deterministic engine assigned each model — rendered from the
+// per-task `evidence` (top-3 candidates) the regenerate / generate-policy endpoints return. The first
+// candidate is the chosen winner; the others are the runners-up it was ranked against. Shared by the
+// global routing policy editor and the per-application policy editor.
+export default function ReasoningDrawer({ evidence, catalog = [], goal, generatorModel, onClose }) {
+  const labelOf = {};
+  for (const t of catalog) labelOf[t.id] = t.label;
+  // Order: catalog order first (only those present in evidence), then any custom task types after.
+  const catalogIds = catalog.map((t) => t.id).filter((id) => evidence[id]);
+  const rest = Object.keys(evidence).filter((id) => !catalogIds.includes(id));
+  const taskIds = [...catalogIds, ...rest];
+  // If every winning pick is a non-benchmarked (estimated) model, quality scores are close and cost
+  // decides — which is exactly why one cheap model can sweep every task. Surface that up front.
+  const allEstimated = taskIds.length > 0 && taskIds.every((id) => evidence[id]?.[0]?.measured === false);
+
+  return (
+    <Drawer title="How the AI chose each model" onClose={onClose}>
+      <p className="text-sm text-gray-500">
+        For each task the engine gates candidates on the capabilities the task needs, scores their
+        task-weighted quality, prices each at its expected cost on your recent traffic, then — for the{" "}
+        <span className="font-medium">{goal}</span> goal — picks the{" "}
+        {goal === "quality" ? "highest-quality model" : "cheapest model that clears the quality bar"}.
+        {generatorModel ? <> Generated by <span className="font-mono text-xs">{generatorModel}</span>.</> : null}
+      </p>
+      {allEstimated && (
+        <div className="rounded-lg border border-amber-200 bg-amber-50 p-3 text-xs text-amber-800">
+          None of these models have benchmark data (LiveBench / LMSYS) yet, so their quality is{" "}
+          <span className="font-medium">estimated</span> and their scores land close together. When quality ties, the
+          cheapest model wins every task — which is why one model can be assigned across the board. Sync model
+          benchmarks or run shadow evals to differentiate them.
+        </div>
+      )}
+      <div className="space-y-3">
+        {taskIds.map((id) => {
+          const cands = evidence[id] || [];
+          const winner = cands[0];
+          return (
+            <div key={id} className="overflow-hidden rounded-lg border border-gray-200">
+              <div className="flex items-center justify-between border-b border-gray-100 px-4 py-2.5">
+                <div className="min-w-0">
+                  <span className="text-sm font-medium text-arbr-charcoal">{labelOf[id] || id}</span>
+                  {labelOf[id] && <span className="ml-2 font-mono text-xs text-gray-400">{id}</span>}
+                </div>
+                {winner && <Badge tone={CONFIDENCE_TONE[winner.confidence] || "gray"}>{winner.confidence} confidence</Badge>}
+              </div>
+              {cands.length === 0 ? (
+                <p className="px-4 py-3 text-xs text-gray-400">No eligible candidates — used the default model.</p>
+              ) : (
+                <table className="w-full text-sm">
+                  <thead>
+                    <tr className="bg-gray-50 text-left text-xs uppercase tracking-wide text-gray-400">
+                      <th className="px-4 py-2 font-medium">Model</th>
+                      <th className="px-4 py-2 font-medium">Quality</th>
+                      <th className="px-4 py-2 font-medium">Est. cost / 1k</th>
+                      <th className="px-4 py-2 font-medium">Data</th>
+                    </tr>
+                  </thead>
+                  <tbody>
+                    {cands.map((c, i) => (
+                      <tr key={c.model} className={`border-t border-gray-100 ${i === 0 ? "bg-green-50/50" : ""}`}>
+                        <td className="px-4 py-2">
+                          <div className="flex items-center gap-2">
+                            <span className="font-mono text-xs text-arbr-charcoal">{c.model} <span className="text-gray-400">({c.tier})</span></span>
+                            {i === 0 && <Badge tone="green">chosen</Badge>}
+                          </div>
+                        </td>
+                        <td className="px-4 py-2 tabular-nums text-gray-600">{c.quality != null ? c.quality.toFixed(2) : "—"}</td>
+                        <td className="px-4 py-2 tabular-nums text-gray-600">{c.expectedCostPer1k != null ? fmt.money(c.expectedCostPer1k) : "—"}</td>
+                        <td className="px-4 py-2 text-xs">
+                          {c.measured
+                            ? <span className="text-gray-500">measured</span>
+                            : <span className="text-amber-600" title="Inferred from a curated table or the model name — no benchmark score">estimated</span>}
+                        </td>
+                      </tr>
+                    ))}
+                  </tbody>
+                </table>
+              )}
+              {winner?.reason && <p className="border-t border-gray-100 px-4 py-2 text-xs text-gray-500">Why: {winner.reason}</p>}
+            </div>
+          );
+        })}
+      </div>
+    </Drawer>
+  );
+}
